# docs/swagger.yaml
basePath: /api
definitions:
...
  auth.LoginResponse:
    properties:
      avatar_url:
        example: users/avatars/fjaun99f-8fna-h8ff-afvd-lmc01mca9jca.png
        type: string
      csrf_token:
        type: string
      email:
        example: ivan@example.com
        type: string
      name:
        example: Иван
        type: string
    type: object
  auth.RegisterRequest:
    properties:
      email:
        example: user@mail.ru
        type: string
      name:
        example: Иван
        type: string
      password:
        example: qwerty12345
        maxLength: 128
        minLength: 8
        type: string
    type: object
  auth.RegisterResponse:
    properties:
      created_at:
        example: 2006-01-02T15:04:05Z07:00
        type: string
      csrf_token:
        type: string
      email:
        example: user@mail.ru
        type: string
      name:
        example: Иван
        type: string
    type: object
...
  github_com_go-park-mail-ru_2026_1_NaNcats_api-gateway_internal_delivery_http_auth.CSRFResponse:
    properties:
      csrf_token:
        type: string
      message:
        type: string
    type: object
  github_com_go-park-mail-ru_2026_1_NaNcats_api-gateway_internal_delivery_http_auth.LoginRequest:
    properties:
      login:
        example: user@mail.ru
        type: string
      password:
        example: qwerty12345
        maxLength: 128
        minLength: 8
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: Неверный формат запроса
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API бэкенда для проекта "Delivery Club" от команды NaNcats
  title: NaNcats Delivery API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Проверяет учетные данные и устанавливает сессионную куку
      parameters:
      - description: Данные для входа
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_NaNcats_api-gateway_internal_delivery_http_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный вход
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Неверный формат JSON
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Неверный логин или пароль
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Авторизация пользователя
      tags:
      - auth
...
  /auth/me:
    get:
      consumes:
      - application/json
      description: Возвращает данные профиля пользователя, если сессионная кука валидна
      produces:
      - application/json
      responses:
        "200":
          description: Успешный вход и создание сессии
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Проверка текущей сессии
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Проверяет данные, создает нового пользователя и устанавливает сессионную
        куку
      parameters:
      - description: Данные для регистрации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешная регистрация
          schema:
            $ref: '#/definitions/auth.RegisterResponse'
        "400":
          description: Ошибка валидации (email/пароль)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Пользователь с такой почтой уже существует
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Регистрация пользователя
      tags:
      - auth
...
  /csrf:
    get:
      description: Возвращает текущий CSRF токен пользователя на основе его сессии.
        Если сессия не найдена, возвращает соответствующее сообщение.
      produces:
      - application/json
      responses:
        "200":
          description: Успешное получение токена или сообщение об отсутствии сессии
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_NaNcats_api-gateway_internal_delivery_http_auth.CSRFResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Получение CSRF токена
      tags:
      - auth
...
swagger: "2.0"